fix: Add "Contract" suffix for consistency
