+++ packages/protocol-kit/src/adapters/web3/contracts/SafeProxyFactory/v1.4.1/SafeProxyFactoryContract_v1_4_1_Web3.ts
@@ -24,21 +24,21 @@ export interface CreateProxyProps extends CreateProxyPropsGeneral {
 }
 
 /**
- * SafeProxyFactory_v1_4_1_Web3 is the implementation specific to the Safe Proxy Factory contract version 1.4.1.
+ * SafeProxyFactoryContract_v1_4_1_Web3 is the implementation specific to the Safe Proxy Factory contract version 1.4.1.
  *
  * This class specializes in handling interactions with the Safe Proxy Factory contract version 1.4.1 using Web3.js.
  *
  * @extends SafeProxyFactoryBaseContractWeb3<SafeProxyFactoryContract_v1_4_1_Abi> - Inherits from SafeProxyFactoryBaseContractWeb3 with ABI specific to Safe Proxy Factory contract version 1.4.1.
  * @implements SafeProxyFactoryContract_v1_4_1_Contract - Implements the interface specific to Safe Proxy Factory contract version 1.4.1.
  */
-class SafeProxyFactory_v1_4_1_Web3
+class SafeProxyFactoryContract_v1_4_1_Web3
   extends SafeProxyFactoryBaseContractWeb3<DeepWriteable<SafeProxyFactoryContract_v1_4_1_Abi>>
   implements SafeProxyFactoryContract_v1_4_1_Contract
 {
   safeVersion: SafeVersion
 
   /**
-   * Constructs an instance of SafeProxyFactory_v1_4_1_Web3
+   * Constructs an instance of SafeProxyFactoryContract_v1_4_1_Web3
    *
    * @param chainId - The chain ID where the contract resides.
    * @param web3Adapter - An instance of Web3Adapter.
@@ -181,4 +181,4 @@ class SafeProxyFactory_v1_4_1_Web3
   }
 }
 
-export default SafeProxyFactory_v1_4_1_Web3
+export default SafeProxyFactoryContract_v1_4_1_Web3
