+++ packages/protocol-kit/src/adapters/web3/contracts/contractInstancesWeb3.ts
@@ -3,10 +3,10 @@ import SafeContract_v1_1_1_Web3 from '@safe-global/protocol-kit/adapters/web3/co
 import SafeContract_v1_2_0_Web3 from '@safe-global/protocol-kit/adapters/web3/contracts/Safe/v1.2.0/SafeContract_v1_2_0_Web3'
 import SafeContract_v1_3_0_Web3 from '@safe-global/protocol-kit/adapters/web3/contracts/Safe/v1.3.0/SafeContract_v1_3_0_Web3'
 import SafeContract_v1_4_1_Web3 from '@safe-global/protocol-kit/adapters/web3/contracts/Safe/v1.4.1/SafeContract_v1_4_1_Web3'
-import SafeProxyFactory_v1_0_0_Web3 from '@safe-global/protocol-kit/adapters/web3/contracts/SafeProxyFactory/v1.0.0/SafeProxyFactory_v1_0_0_Web3'
-import SafeProxyFactory_v1_1_1_Web3 from '@safe-global/protocol-kit/adapters/web3/contracts/SafeProxyFactory/v1.1.1/SafeProxyFactory_v1_1_1_Web3'
-import SafeProxyFactory_v1_3_0_Web3 from '@safe-global/protocol-kit/adapters/web3/contracts/SafeProxyFactory/v1.3.0/SafeProxyFactory_v1_3_0_Web3'
-import SafeProxyFactory_v1_4_1_Web3 from '@safe-global/protocol-kit/adapters/web3/contracts/SafeProxyFactory/v1.4.1/SafeProxyFactory_v1_4_1_Web3'
+import SafeProxyFactoryContract_v1_0_0_Web3 from '@safe-global/protocol-kit/adapters/web3/contracts/SafeProxyFactory/v1.0.0/SafeProxyFactoryContract_v1_0_0_Web3'
+import SafeProxyFactoryContract_v1_1_1_Web3 from '@safe-global/protocol-kit/adapters/web3/contracts/SafeProxyFactory/v1.1.1/SafeProxyFactoryContract_v1_1_1_Web3'
+import SafeProxyFactoryContract_v1_3_0_Web3 from '@safe-global/protocol-kit/adapters/web3/contracts/SafeProxyFactory/v1.3.0/SafeProxyFactoryContract_v1_3_0_Web3'
+import SafeProxyFactoryContract_v1_4_1_Web3 from '@safe-global/protocol-kit/adapters/web3/contracts/SafeProxyFactory/v1.4.1/SafeProxyFactoryContract_v1_4_1_Web3'
 import Web3Adapter from '@safe-global/protocol-kit/adapters/web3/Web3Adapter'
 import { SafeContract_v1_1_1_Abi } from '@safe-global/protocol-kit/contracts/AbiType/Safe/v1.1.1/SafeContract_v1_1_1'
 import { SafeContract_v1_2_0_Abi } from '@safe-global/protocol-kit/contracts/AbiType/Safe/v1.2.0/SafeContract_v1_2_0'
@@ -220,7 +220,7 @@ export async function getSafeProxyFactoryContractInstance(
 
   switch (safeVersion) {
     case '1.4.1':
-      safeProxyFactoryContract = new SafeProxyFactory_v1_4_1_Web3(
+      safeProxyFactoryContract = new SafeProxyFactoryContract_v1_4_1_Web3(
         chainId,
         web3Adapter,
         contractAddress,
@@ -229,7 +229,7 @@ export async function getSafeProxyFactoryContractInstance(
       )
       return safeProxyFactoryContract.mapToTypechainContract() // remove this mapper after remove typechain
     case '1.3.0':
-      safeProxyFactoryContract = new SafeProxyFactory_v1_3_0_Web3(
+      safeProxyFactoryContract = new SafeProxyFactoryContract_v1_3_0_Web3(
         chainId,
         web3Adapter,
         contractAddress,
@@ -239,7 +239,7 @@ export async function getSafeProxyFactoryContractInstance(
       return safeProxyFactoryContract.mapToTypechainContract() // remove this mapper after remove typechain
     case '1.2.0':
     case '1.1.1':
-      safeProxyFactoryContract = new SafeProxyFactory_v1_1_1_Web3(
+      safeProxyFactoryContract = new SafeProxyFactoryContract_v1_1_1_Web3(
         chainId,
         web3Adapter,
         contractAddress,
@@ -248,7 +248,7 @@ export async function getSafeProxyFactoryContractInstance(
       )
       return safeProxyFactoryContract.mapToTypechainContract() // remove this mapper after remove typechain
     case '1.0.0':
-      safeProxyFactoryContract = new SafeProxyFactory_v1_0_0_Web3(
+      safeProxyFactoryContract = new SafeProxyFactoryContract_v1_0_0_Web3(
         chainId,
         web3Adapter,
         contractAddress,
